requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
defusedxml==0.7.0rc1
Django==3.0.7
django-allauth==0.43.0
django-autocomplete-light==3.5.1
django-cleanup==5.1.0
django-colorfield==0.3.0
django-crispy-forms==1.9.1
django-embed-video==1.3.2
django-faicon==0.1.0
django-smart-selects==1.5.6
gunicorn==20.0.4
idna==2.9
oauthlib==3.1.0
Pillow==7.1.2
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==3.4.0
social-auth-core==3.3.3
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.2.0
